move all widgets to new queries, update glad widget to use minimal changes
