+++ app/javascript/components/widgets/widgets/forest-change/glad-alerts/actions.js
@@ -5,12 +5,12 @@ export default ({ params }) =>
   axios.all([fetchGladAlerts(params), fetchGLADLatest(params)]).then(
     axios.spread((alerts, latest) => {
       let data = {};
-      if (alerts && latest) {
+      if (alerts && alerts.data && latest) {
         const latestDate =
           latest && latest.attributes && latest.attributes.updatedAt;
 
         data = {
-          alerts,
+          alerts: alerts.data.data,
           latest: latestDate,
           settings: { latestDate }
         };
