+++ app/javascript/components/widgets/widgets/forest-change/glad-ranked/actions.js
@@ -1,5 +1,5 @@
 import { fetchGladIntersectionAlerts, fetchGLADLatest } from 'services/alerts';
-import { getExtentGrouped } from 'services/forest-data';
+import { getExtentGrouped } from 'services/analysis-cached';
 import axios from 'axios';
 
 export default ({ params }) =>
