+++ app/javascript/components/widgets/widgets/forest-change/tree-loss-plantations/actions.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { getLoss } from 'services/forest-data';
+import { getLoss } from 'services/analysis-cached';
 
 export default ({ params }) =>
   axios
