+++ app/javascript/components/widgets/widgets/land-cover/tree-cover-located/actions.js
@@ -1,4 +1,4 @@
-import { getExtentGrouped } from 'services/forest-data';
+import { getExtentGrouped } from 'services/analysis-cached';
 
 export default ({ params }) =>
   getExtentGrouped(params).then(response => {
