+++ app/javascript/components/widgets/widgets/land-cover/tree-cover/actions.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { getExtent } from 'services/forest-data';
+import { getExtent } from 'services/analysis-cached';
 
 export default ({ params }) =>
   axios
